feat: rename old foods app to foods_old
