+++ caloru/caloru/urls.py
@@ -23,7 +23,7 @@ from . import views
 
 urlpatterns = [
     path("", views.home, name="home"),
-    path("foods/", include("foods.urls")),
+    path("foods/", include("foods_old.urls")),
     path("admin", admin.site.urls),
     path("accounts/", include("accounts.urls")),
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
