+++ caloru/foods/admin.py
@@ -1,7 +1,3 @@
 from django.contrib import admin
 
-from .models import ConsumedProduct, Product
-
 # Register your models here.
-admin.site.register(Product)
-admin.site.register(ConsumedProduct)
