+++ caloru/foods_old/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FoodsOldConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "foods_old"
