+++ pyproject.toml
@@ -7,6 +7,7 @@ requires-python = ">=3.13"
 dependencies = [
     "dj-database-url>=3.0.1",
     "django>=5.2.6",
+    "django-rename-app>=0.1.7",
     "django-widget-tweaks>=1.5.0",
     "djangorestframework>=3.16.1",
     "uvicorn>=0.35.0",
